Refactor clip property handling and UI integration
- Updated Project and Timeline to use ClipBindingValue instead of PropertyValue for clip properties.
- Removed unused code related to nested timelines in Project.
- Simplified ClipData and PropertyValue structures.
- Enhanced video rendering code by removing unnecessary imports.
- Added new FFI functions for clip property management, including getting and setting clip properties.
- Introduced a new ClipPropertiesPanel in the GUI for displaying and editing clip properties with collapsible sections.
- Implemented FieldControlFactory to create UI controls for various property types.
- Improved error handling in result types across the project.

+++ lib/src/util/color.rs
@@ -14,8 +14,8 @@ use rkyv::{Archive, CheckBytes, bytecheck};
 #[archive_attr(derive(CheckBytes))]
 #[repr(C)]
 pub struct RgbaColor {
-    pub r: u8,
-    pub g: u8,
-    pub b: u8,
-    pub a: u8,
+    pub r: f32,
+    pub g: f32,
+    pub b: f32,
+    pub a: f32,
 }
